Move introductions + add simple jobs validation
