--- bot/cogs/handlers.py
@@ -1,5 +1,0 @@
-import discord
-
-from .variables import *
-
-# TODO: add message embed handler
